+++ CHANGES.rst
@@ -83,6 +83,13 @@ Bug fixes:
 
   __ https://github.com/atviriduomenys/spinta/issues/626
 
+Bug fixes:
+
+- Added support for `getone` for `sql` backend (`#513`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/513
+
+
 0.1.62 (2024-02-29)
 ===================